Uploading Readmen.md and Jupyter
This repository contains the final assessment of module 1, bootcamp Data Analysis, group K for Fiona Sánchez Rafferty.
The files in the repository are a Jupyter containing the code and a Readme.md outlining the committed changes.
The assessment consists in generating code on python to create a class called TiendaOnline with specific attributes, methods and instances.